yolox/requirements.txt
requests == 2.28.2
opencv_python==4.8.1.78
kiwisolver==1.4.5
Pillow==8.4.0
numpy==1.25.2
torch==2.0.0a0
MarkupSafe==2.1.3
torchvision==0.15.1a0
psutil==5.9.6
pylibtiff==0.6.1
pytesseract==0.3.10
ninja==1.11.1
loguru==0.7.2
tqdm==4.66.1
thop==0.1.1.post2209072238
pycocotools==2.0.7
matplotlib==3.8.1
yolox==0.3.0